submit-project: add link to projects repository
Move the form to a dedicated client component
